페이지 선택시 내용을 전달하는 부분 strategy로 이동, 초기화 완료후 0페이지를 발행하던 부분 view 내부로 이동
