+++ app/src/main/java/net/jspiner/epub_viewer/ui/reader/ReaderActivity.kt
@@ -92,7 +92,6 @@ class ReaderActivity : BaseActivity<ActivityReaderBinding, ReaderViewModel>() {
     private fun calculatePage() {
         getPaginator(baseContext, viewModel.extractedEpub).calculatePage()
             .doOnSuccess { viewModel.setPageInfo(it) }
-            .doOnSuccess { viewModel.onPagerItemSelected(0) }
             .subscribeOn(Schedulers.computation())
             .observeOn(AndroidSchedulers.mainThread())
             .doOnSubscribe { showLoading() }
